Add types to base
